+++ bootstrapvz/providers/azure/__init__.py
@@ -1,10 +1,10 @@
 from bootstrapvz.common import task_groups
 import tasks.packages
 import tasks.boot
-import tasks.image
 from bootstrapvz.common.tasks import loopback
 from bootstrapvz.common.tasks import initd
 from bootstrapvz.common.tasks import ssh
+from bootstrapvz.common.tasks import apt
 
 
 def validate_manifest(data, validator, error):
@@ -15,17 +15,18 @@ def validate_manifest(data, validator, error):
 
 def resolve_tasks(taskset, manifest):
 	taskset.update(task_groups.get_standard_groups(manifest))
-
-	taskset.update([tasks.packages.DefaultPackages,
+	taskset.update([apt.AddBackports,
+	                tasks.packages.DefaultPackages,
 	                loopback.AddRequiredCommands,
 	                loopback.Create,
+	                loopback.MoveImage,
 	                initd.InstallInitScripts,
 	                ssh.AddOpenSSHPackage,
 	                ssh.ShredHostkeys,
 	                ssh.AddSSHKeyGeneration,
 	                tasks.packages.Waagent,
 	                tasks.boot.ConfigureGrub,
-	                tasks.image.ConvertToVhd,
+	                tasks.boot.PatchUdev,
 	                ])