+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+2015-06-29
+----------
+Alex Adriaanse:
+	* Fix DKMS kernel version error
+	* Add support for Btrfs
+	* Add EC2 Jessie HVM manifest
+
+2015-05-08
+----------
+Alexandre Derumier:
+    * Fix #219: ^PermitRootLogin regex
+
 2015-05-02
 ----------
 Anders Ingemann: